Split distribution operator test cases
