--- tests/test_distribution_ops.py
@@ -1,178 +1,0 @@
-import numpy as np
-import pytest
-import scipy
-import torch
-
-import flag_gems
-
-from .accuracy_utils import DISTRIBUTION_SHAPES, FLOAT_DTYPES, to_reference
-
-device = flag_gems.device
-
-
-@pytest.mark.normal
-@pytest.mark.parametrize(
-    "op_name, float_t, scale_t",
-    [
-        pytest.param(
-            name,
-            float,
-            scale,
-            marks=getattr(pytest.mark, name, None),
-        )
-        for name, float, scale in [
-            ("normal_float_tensor", "mean", "std"),
-            ("normal_tensor_float", "std", "mean"),
-            ("normal_tnsor_tensor", "std", "std"),
-        ]
-    ],
-)
-@pytest.mark.parametrize("shape", DISTRIBUTION_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_normal(op_name, float_t, scale_t, shape, dtype):
-    if flag_gems.vendor_name == "cambricon":
-        torch.manual_seed(42)
-        torch.mlu.manual_seed_all(42)
-    if flag_gems.vendor_name in ["metax", "iluvatar", "kunlunxin"]:
-        torch.manual_seed(42)
-        torch.cuda.manual_seed_all(42)
-    loc = (
-        3.0
-        if float_t == "mean"
-        else torch.full(
-            size=shape, fill_value=3.0, dtype=dtype, device=flag_gems.device
-        )
-    )
-    scale = (
-        10.0
-        if scale_t == "mean"
-        else torch.full(
-            size=shape, fill_value=10.0, dtype=dtype, device=flag_gems.device
-        )
-    )
-    with flag_gems.use_gems():
-        res_out = torch.normal(loc, scale)
-    ref_out = to_reference(res_out)
-    mean = torch.mean(ref_out)
-    std = torch.std(ref_out)
-    assert torch.abs(mean - 3.0) < 0.1
-    assert torch.abs(std - 10.0) < 0.1
-
-
-@pytest.mark.inplace
-@pytest.mark.normal_
-@pytest.mark.parametrize("shape", DISTRIBUTION_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_normal_(shape, dtype):
-    if flag_gems.vendor_name == "cambricon":
-        torch.manual_seed(42)
-        torch.mlu.manual_seed_all(42)
-    if flag_gems.vendor_name in ["metax", "iluvatar"]:
-        torch.manual_seed(42)
-        torch.cuda.manual_seed_all(42)
-    loc = 3.0
-    scale = 10.0
-    res_out = torch.randn(size=shape, dtype=dtype, device=flag_gems.device)
-    with flag_gems.use_gems():
-        res_out.normal_(loc, scale)
-    ref_out = to_reference(res_out)
-    mean = torch.mean(ref_out)
-    std = torch.std(ref_out)
-    assert torch.abs(mean - 3.0) < 0.1
-    assert torch.abs(std - 10.0) < 0.1
-
-
-@pytest.mark.uniform_
-@pytest.mark.parametrize("shape", DISTRIBUTION_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_uniform(shape, dtype):
-    x = torch.randn(size=shape, dtype=dtype, device=flag_gems.device)
-    with flag_gems.use_gems():
-        x.uniform_(-3, 3)
-    assert (x <= 3.0).all()
-    assert (x >= -3.0).all()
-
-
-@pytest.mark.exponential_
-@pytest.mark.parametrize("shape", DISTRIBUTION_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_exponential_(shape, dtype):
-    x = torch.empty(size=shape, dtype=dtype, device=flag_gems.device)
-    with flag_gems.use_gems():
-        x.exponential_()
-    assert x.min() > 0
-
-
-@pytest.mark.exponential_
-@pytest.mark.parametrize("shape", DISTRIBUTION_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_fast_exponential_(shape, dtype):
-    x = torch.empty(size=shape, dtype=dtype, device=flag_gems.device)
-    lambd = 1.0
-    mean_tol = 0.05
-    var_tol = 0.05
-    with flag_gems.use_gems():
-        x.exponential_()
-    x_res = to_reference(x)
-    mean_res = torch.mean(x_res.to(torch.float32)).to(dtype)
-    var_res = torch.var(x_res.to(torch.float32)).to(dtype)
-    mean_ref = 1.0 / lambd
-    var_ref = 1.0 / (lambd**2)
-    assert torch.abs(mean_res - mean_ref) < mean_tol
-    assert torch.abs(var_res - var_ref) < var_tol
-
-
-@pytest.mark.bernoulli_
-@pytest.mark.parametrize("shape", DISTRIBUTION_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_bernoulli_(shape, dtype):
-    x = torch.empty(size=shape, dtype=dtype, device=flag_gems.device)
-    p = 0.5
-    with flag_gems.use_gems():
-        x.bernoulli_(p)
-    # Check that all values are 0 or 1
-    assert ((x == 0) | (x == 1)).all()
-    # Check that the mean is approximately p (statistical test)
-    mean = x.float().mean().item()
-    assert abs(mean - p) < 0.1
-
-
-@pytest.mark.bernoulli_
-@pytest.mark.parametrize("shape", DISTRIBUTION_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-@pytest.mark.parametrize("p", [0.0, 0.3, 0.7, 1.0])
-def test_accuracy_bernoulli_various_p(shape, dtype, p):
-    x = torch.empty(size=shape, dtype=dtype, device=flag_gems.device)
-    with flag_gems.use_gems():
-        x.bernoulli_(p)
-    # Check that all values are 0 or 1
-    assert ((x == 0) | (x == 1)).all()
-    # Check boundary cases
-    if p == 0.0:
-        assert (x == 0).all()
-    elif p == 1.0:
-        assert (x == 1).all()
-    else:
-        # Check that the mean is approximately p
-        mean = x.float().mean().item()
-        assert abs(mean - p) < 0.15
-
-
-@pytest.mark.multinomial
-@pytest.mark.parametrize("shape", [(1024, 10)])
-@pytest.mark.parametrize("dtype", [torch.float16, torch.float32])
-@pytest.mark.parametrize("n_samples", [2048])
-def test_accuracy_multinomial_with_replacement(shape, dtype, n_samples):
-    # First use multinomial to generate a series of indices, then
-    # use the index counts as the input probabilities (scaled)
-    rand_indices = torch.multinomial(torch.rand(shape), n_samples, True).to(device)
-    inp_counts = torch.nn.functional.one_hot(rand_indices).sum(1)
-    with flag_gems.use_gems():
-        out_indices = torch.multinomial(inp_counts.to(dtype=dtype), n_samples, True)
-    out_counts = torch.nn.functional.one_hot(out_indices).sum(1)
-    # Do a simple Chi-square test
-    assert torch.equal(inp_counts.sum(-1), out_counts.sum(-1))
-    chi2, pvalue = scipy.stats.chisquare(
-        out_counts.tolist(), inp_counts.tolist(), axis=-1
-    )
-    assert np.sum(pvalue < 0.05) / len(pvalue) < 0.1
