+++ tests/test_exponential.py
@@ -1,0 +1,38 @@
+import pytest
+import torch
+
+import flag_gems
+
+from . import accuracy_utils as utils
+
+
+@pytest.mark.exponential_
+@pytest.mark.parametrize("shape", utils.DISTRIBUTION_SHAPES)
+@pytest.mark.parametrize("dtype", utils.FLOAT_DTYPES)
+def test_exponential_(shape, dtype):
+    x = torch.empty(size=shape, dtype=dtype, device=flag_gems.device)
+    with flag_gems.use_gems():
+        x.exponential_()
+
+    assert x.min() > 0
+
+
+@pytest.mark.exponential_
+@pytest.mark.parametrize("shape", utils.DISTRIBUTION_SHAPES)
+@pytest.mark.parametrize("dtype", utils.FLOAT_DTYPES)
+def test_accuracy_fast_exponential_(shape, dtype):
+    x = torch.empty(size=shape, dtype=dtype, device=flag_gems.device)
+    lambd = 1.0
+    mean_tol = 0.05
+    var_tol = 0.05
+    with flag_gems.use_gems():
+        x.exponential_()
+
+    x_res = utils.to_reference(x)
+    mean_res = torch.mean(x_res.to(torch.float32)).to(dtype)
+    var_res = torch.var(x_res.to(torch.float32)).to(dtype)
+    mean_ref = 1.0 / lambd
+    var_ref = 1.0 / (lambd**2)
+
+    assert torch.abs(mean_res - mean_ref) < mean_tol
+    assert torch.abs(var_res - var_ref) < var_tol
